requirements.txt
dash>=2.0.0
requests
flask
dash-bootstrap-components
snowflake-connector
python-dotenv